+++ web/js/ui/cloud_auth/CloudAuthButton.tsx
@@ -28,10 +28,12 @@ export class CloudAuthButton extends React.Component<IProps, IState> {
         let stage: AuthStage | undefined;
 
         if (document.location!.hash === '#login') {
+            RendererAnalytics.event({category: 'cloud', action: 'login'});
             stage = 'login';
         }
 
         if (document.location!.hash === "#configured") {
+            RendererAnalytics.event({category: 'cloud', action: 'configured'});
             stage = 'configured';
         }